pyproject.toml
[project]
name = "mat"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Ethan", email = "xiewenxuan0000@gmail.com" }]
dependencies = [
    "pyside6>=6.7.1",
    "pyopengl>=3.1.7",
    "pyserial>=3.5",
    "numpy>=1.26.4",
    "matplotlib",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
virtual = true
managed = true
dev-dependencies = [
    "pyinstaller>=6.8.0",
    "pandas>=2.2.2",
]
